requirements.txt
click==8.1.8
natsort==8.4.0
numpy==2.1.1
Pillow==11.3.0
pydantic==2.11.1
Requests==2.32.3
pypdfium2==4.30.0
split-image==2.0.1
Flask==3.0.0
Werkzeug==3.0.1
Flask-CORS==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.1
psutil==5.9.8
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
bcrypt==4.1.2
marshmallow==3.20.1
flask-marshmallow==0.15.0
gunicorn==21.2.0
# Optimized API dependencies
celery==5.3.4
redis==5.0.1
flower==2.0.1
prometheus-client==0.19.0
# PostgreSQL driver
psycopg2-binary==2.9.9